WL#13543 InnoDB: Provide partitioning information to RAPID during load
Post-push fix - Fix build issue in 32-bit platforms and a doygen issue.

Approved-by : Kuba Lopuszanski <jakub.lopuszanski@oracle.com>

+++ storage/innobase/include/row0pread.h
@@ -160,7 +160,7 @@ class Parallel_reader {
   */
     Config(const Scan_range &scan_range, dict_index_t *index,
            size_t read_level = 0,
-           size_t partition_id = std::numeric_limits<uint64_t>::max(),
+           size_t partition_id = std::numeric_limits<size_t>::max(),
            bool read_ahead = false)
         : m_scan_range(scan_range),
           m_index(index),
@@ -197,8 +197,8 @@ class Parallel_reader {
     size_t m_read_level{0};
 
     /** Partition id if the index to be scanned belongs to a partitioned table,
-    else std::numeric_limits<uint64_t>::max(). */
-    size_t m_partition_id{std::numeric_limits<uint64_t>::max()};
+    else std::numeric_limits<size_t>::max(). */
+    size_t m_partition_id{std::numeric_limits<size_t>::max()};
 
     /** if true then enable separate read ahead threads. */
     bool m_read_ahead{false};
@@ -709,7 +709,7 @@ class Parallel_reader::Ctx {
   }
 
   /** @return the partition id of the index.
-  @note this is std::numeric_limits<uint64_t>::max() if the index does not
+  @note this is std::numeric_limits<size_t>::max() if the index does not
   belong to a partition. */
   size_t partition_id() const MY_ATTRIBUTE((warn_unused_result)) {
     return m_scan_ctx->m_config.m_partition_id;
